+++ src/content/data.js
@@ -1,4 +1,4 @@
-import { TUICI18N } from "./i18n.js";
+import { TUICI18N } from "./i18n.ts";
 import { TUICLibrary, TUICPref } from "./library.js";
 import { SIDEBAR_BUTTON_ICON } from "./data/icons.js";
