+++ src/content/index.js
@@ -5,7 +5,7 @@
 
 import { TUICObserver } from "./observer.js";
 import { TUICLibrary } from "./library.js";
-import { TUICI18N } from "./i18n.js";
+import { TUICI18N } from "./i18n.ts";
 import { applySystemCss, addCssElement, applyDataCss } from "./applyCSS.js";
 import { isSafemode, runSafemode } from "./safemode.js";
