+++ src/content/observer.js
@@ -4,7 +4,7 @@ import TWITTER from "./icons/logo/twitter.svg";
 import X from "./icons/logo/x.svg";
 import EMPTY from "./icons/logo/empty.svg";
 import { HOME_ICON, SIDEBAR_BUTTON_ICON } from "./data/icons.js";
-import { TUICI18N } from "./i18n.js";
+import { TUICI18N } from "./i18n.ts";
 import { TUICLibrary, TUICPref } from "./library.js";
 import { TUICOptionHTML } from "./option.js";
 import { isSafemode } from "./safemode.js";
