+++ src/content/option.js
@@ -1,6 +1,6 @@
 import { applyCustomCss, applySystemCss } from "./applyCSS.js";
 import { TUICData } from "./data.js";
-import { TUICI18N } from "./i18n.js";
+import { TUICI18N } from "./i18n.ts";
 import { TUICLibrary, TUICPref } from "./library.js";
 import { TUICObserver } from "./observer.js";
 import { isSafemode } from "./safemode.js";
