+++ src/content/safemode.js
@@ -1,4 +1,4 @@
-import { TUICI18N } from "./i18n.js";
+import { TUICI18N } from "./i18n.ts";
 import { TUICOptionHTML } from "./option.js";
 
 export const isSafemode = location.pathname === "/tuic/safemode";
