+++ src/options/options.html
@@ -3,7 +3,7 @@
     <head>
         <title class="i18n" i18n-id="optionTitle">設定及び操作方法(Twitter UI Customizer)</title>
         <meta charset="utf-8" />
-        <script type="module" src="../content/i18n.js"></script>
+        <script type="module" src="../content/i18n.ts"></script>
         <script type="module" src="./options.js"></script>
     </head>
     <body>
