+++ src/popup/popup.html
@@ -6,7 +6,7 @@
         <link rel="stylesheet" href="./popup.scss" />
     </head>
     <body>
-        <script type="module" src="../content/i18n.js"></script>
+        <script type="module" src="../content/i18n.ts"></script>
         <script type="module" src="./popup.ts"></script>
         <img id="logo" src="../../icon/TUIC_TextWhite.svg" />
         <div class="w-72">
