+++ src/safemode.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html>
+<html lang="ja">
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
